+++ pkgs/jnigen/lib/src/generate_bindings.dart
@@ -11,6 +11,7 @@ import 'bindings/excluder.dart';
 import 'bindings/kotlin_processor.dart';
 import 'bindings/linker.dart';
 import 'bindings/renamer.dart';
+import 'bindings/stub_collector.dart';
 import 'bindings/visitor.dart';
 import 'config/config.dart';
 import 'elements/elements.dart';
@@ -46,7 +47,7 @@ Future<void> generateJniBindings(Config config) async {
   // Keep the order in sync with `elements/elements.dart`.
   var stage = GenerationStage.userVisitors;
   R runStage<R>(TopLevelVisitor<R> visitor) {
-    assert(visitor.stage.index == stage.index + 1);
+    assert(visitor.stage.index > stage.index);
     stage = visitor.stage;
     return classes.accept(visitor);
   }
@@ -54,6 +55,9 @@ Future<void> generateJniBindings(Config config) async {
   runStage(Excluder(config));
   runStage(KotlinProcessor());
   await runStage(Linker(config));
+  if (config.generateStubs) {
+    runStage(StubCollector());
+  }
   runStage(Renamer(config));
   // classes.accept(const Printer());