+++ pkgs/jnigen/test/stub_test/java/com/example/A.java
@@ -1,0 +1,6 @@
+package com.example;
+public class A {
+  public B getB() { return new B(); }
+  public void takeB(B b) {}
+  public void takeD(D d) {}
+}