+++ pkgs/jnigen/test/stub_test/java/com/example/B.java
@@ -1,0 +1,4 @@
+package com.example;
+public class B {
+  public E e;
+}