+++ pkgs/jnigen/test/stub_test/java/com/example/C.java
@@ -1,0 +1,2 @@
+package com.example;
+public class C extends D {}