+++ pkgs/jnigen/test/stub_test/java/com/example/D.java
@@ -1,0 +1,2 @@
+package com.example;
+public class D extends E {}