+++ pkgs/jnigen/test/stub_test/java/com/example/E.java
@@ -1,0 +1,2 @@
+package com.example;
+public class E {}